Fix 'fetch' not being able to find an item at all
So, this fixes a fairly peculiar issue with the `fetch` method.

The `const data = await this.fetchAll(config);` method returns an `Array`, and the old code used `hasOwnProperty`, which is meant for objects. That meant the code would always return false, and whatever the engine was looking for would never be found.

After a bit of brainstorming with Sean, we figured it out. I changed it to a `find` method, instead, and it works like a charm.